RedBusAutomationFrameWork/requirements.txt
These are the specified libraries which are requred to execute this project


Python 				3.7x	3.7x

PyHamcrest			1.9.0	1.9.0

html-testRunner			1.2	1.2

openpyxl			2.6.2	2.6.2

pip				19.1.1	19.1.1

py				1.8.0	1.8.0

pyparsing			2.4.0	2.4.0

pytest				5.0.1	5.0.1

pytest-forked			1.0.2	1.0.2

pytest-html			1.21.1	1.21.1

pytest-metadata			1.8.0	1.8.0

pytest-xdist			1.29.0	1.29.0

selenium			3.141.0	3.141.0

